fix: Address PR #1396 review feedback (mouse scroll)
- Add separate thread_mouse_scroll_accumulator so thread panel scroll
  credits don't bleed into main chat (and vice versa)
- Move 4 inline scroll tests from app.rs mod tests to mouse_scroll_tests.rs
  per CLAUDE.md convention; add cross-panel isolation test
- Add width.max(1) guard to count_line_wraps to prevent infinite loop
  when called with width=0 (matches existing guard in wrap_line)
- Add equivalence tests: count_wrapped_lines must match wrap_content().len()
  across typical inputs, multiline text, and the zero-width edge case
- Reset thread_scroll_offset in load_channel_messages so switching
  channels doesn't leave the thread panel at a stale scroll position
